Blacklist br_netfilter
br_netfilter turns on iptables filtering for network bridges. That
completely breaks Whonix (took me long enough to figure that out), and
doesn't really make much sense anyways unless I'm misunderstanding
something.

The other option is to change parameters in sysctl, but they seem to be
evaluated before the module is loaded so that's no good.